IIA_cutoff/requirements.txt
# ---------------- Core ----------------
python-dotenv>=1.0.0
numpy>=1.23.0
pandas>=2.0.0

# ---------------- Databases ----------------
pymongo>=4.6.0
psycopg2-binary>=2.9.9
redis>=5.0.1

# ---------------- Machine Learning ----------------
scikit-learn>=1.3.0

# ---------------- Backend ----------------
flask>=3.0.0
flask-cors>=4.0.0

# ---------------- OpenAI (NL → SQL) ----------------
openai>=1.3.5

# ---------------- Utilities ----------------
tqdm>=4.66.0
